KEYCLOAK-13018: find user by id instead of username and fix FindUserByUsername implementation
We are currently finding a user by username to reconcile the KeycloakUser custom resource. However, a username can be mutable so we should retrieve a user by it's ID instead to make sure we're reconciling the correct user. The Keycloak client's generic create function was changed to return the uid of the newly created resource. This uid is now added back into the KeycloakUser custom resource spec after user creation. This way, we can now find the user by id instead of by username.

The FindUserByUsername implementation was also changed. It returns the first user which contains the string specified in the search query. If a KeycloakUser was created or deleted with a username that is a substring of a user that already exists in the realm and is the first result of that query, it will update or delete that user instead. The following changes ensures that the user found by this function returns the user that matches the correct username. The parameter to the request was also changed to filter users by username instead of all the fields.

https://issues.redhat.com/browse/KEYCLOAK-13018